refactor: 리팩토링
ChatMessageSchdeuler 현재 진행중인 스터디 목록 가져와서 해당 스터디들 채팅 내역 정리하도록 변경

투표 생성 시
유권자 수도 redis에 저장
투표 종료 시간 전 투표 참여 수가 유권자 수와 같거나 강퇴 투표의 경우 찬성이 과반수를 넘긴 경우
즉시 투표 집계 후 종료하도록 변경

quatz schduler jobName, groupName 일정 변경 투표, 강퇴 투표
동일하게 변경(jobName : "job" + voteId, groupName : "job")

쓸데없는 jpql 쿼리 메서드로
변경

+++ src/main/java/com/campfiredev/growtogether/study/service/vote/KickVoteProcessor.java
@@ -24,7 +24,7 @@ public class KickVoteProcessor implements VoteProcessor {
 
   @Override
   @Transactional
-  public void processVote(VoteEntity voteEntity, int votes, int totalSize) {
+  public void processVote(VoteEntity voteEntity, Long votes, Long totalSize) {
     if (votes > totalSize / 2) {
       KickVoteEntity kickVoteEntity = (KickVoteEntity) voteEntity;
       StudyMemberEntity studyMemberEntity = joinRepository.findById(kickVoteEntity.getTarget().getId())
